Move dow_number/2 to julian/util
This predicate is not part of the main julian interface so it shouldn't
be exported from library(julian).  However, it's useful enough that
others are likely to want access to it.  These kinds of low-level
predicates are now housed in library(julian/util) where they can be used
independently of library(julian).